moved SQL Initializer
